fix: replaced the custom date parser in the stdlib with a shared one from the compiler
