+++ compiler/plc_lexer/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "plc_lexer"
+version = "1.1.0-dev"
+edition = "2021"
+license = "LGPL-3.0"
+description = "Structured Text lexer and parse session"
+
+[dependencies]
+logos.workspace = true
+plc_ast = { path = "../plc_ast", version = "1.1.0-dev" }
+plc_diagnostics = { path = "../plc_diagnostics", version = "1.1.0-dev" }
+plc_source = { path = "../plc_source", version = "1.1.0-dev" }
+
+[dev-dependencies]
+pretty_assertions = "1.3.0"
