+++ compiler/plc_lexer/src/lib.rs
@@ -1,0 +1,5 @@
+//! Structured Text lexer and parse-session types shared by compiler components.
+
+pub mod lexer;
+
+pub use lexer::{lex, lex_with_ids, ParseSession, Token, TokenClass};
