+++ compiler/plc_parser/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "plc_parser"
+version = "1.1.0-dev"
+edition = "2021"
+license = "LGPL-3.0"
+description = "Shared Structured Text literal parser"
+
+[dependencies]
+plc_ast = { path = "../plc_ast", version = "1.1.0-dev" }
+plc_diagnostics = { path = "../plc_diagnostics", version = "1.1.0-dev" }
+plc_lexer = { path = "../plc_lexer", version = "1.1.0-dev" }
+plc_source = { path = "../plc_source", version = "1.1.0-dev" }
