+++ libs/stdlib/Cargo.toml
@@ -29,9 +29,14 @@ version = "1.1.0-dev"
 
 [dependencies]
 chrono = "0.4"
+plc_ast = { path = "../../compiler/plc_ast", version = "1.1.0-dev" }
+plc_lexer = { path = "../../compiler/plc_lexer", version = "1.1.0-dev" }
+plc_parser = { path = "../../compiler/plc_parser", version = "1.1.0-dev" }
 num = "0.4"
 paste = "1.0.8"
 log = "0.4"
+
+plc_source = { path = "../../compiler/plc_source", version = "1.1.0-dev" }
 
 [dev-dependencies]
 libloading = "0.9"
