+++ src/lib.rs
@@ -37,7 +37,8 @@ pub mod expression_path;
 pub mod hardware_binding;
 pub mod hw_map;
 pub mod index;
-pub mod lexer;
+pub use plc_lexer as lexer;
+pub use plc_lexer::expect_token;
 pub mod linker;
 pub mod lowering;
 pub mod output;
